// src/api/incident-log/models/Incident.ts
export interface Incident {
    /**
     * Unique id of the incident
     * @type {string}
     * @memberof Incident
     */
    id: string;
    /**
     * Type of the incident
     * @type {string}
     * @memberof Incident
     */
    incidentType: string;
    /**
     * Place where the incident occurred
     * @type {string}
     * @memberof Incident
     */
    location: string;
    /**
     * Date and time when the incident occurred
     * @type {Date}
     * @memberof Incident
     */
    occurredAt: Date;
    /**
     * Description of the incident
     * @type {string}
     * @memberof Incident
     */
    description: string;
    /**
     * 
     * @type {IncidentSeverity}
     * @memberof Incident
     */
    severity: IncidentSeverity;
    /**
     * 
     * @type {IncidentStatus}
     * @memberof Incident
     */
    status: IncidentStatus;
    /**
     * List of attached files
     * @type {Array<string>}
     * @memberof Incident
     */
    attachments?: Array<string>;
    /**
     * Investigation report
     * @type {string}
     * @memberof Incident
     */
    investigationReport?: string;
    /**
     * Additional notes
     * @type {string}
     * @memberof Incident
     */
    notes?: string;
}